move feature selection into folder
